# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChatWave API",
    "description": "Real-time chat application API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/friends/request": {
      "post": {
        "tags": [
          "Friends"
        ],
        "summary": "Send Friend Request",
        "description": "Send a friend request to another user.",
        "operationId": "send_friend_request_api_friends_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendshipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/friends/requests": {
      "get": {
        "tags": [
          "Friends"
        ],
        "summary": "Get Friend Requests",
        "description": "Get all pending friend requests for the current user.",
        "operationId": "get_friend_requests_api_friends_requests_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendRequestResponse"
                  },
                  "type": "array",
                  "title": "Response Get Friend Requests Api Friends Requests Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/friends/requests/{request_id}": {
      "put": {
        "tags": [
          "Friends"
        ],
        "summary": "Respond To Friend Request",
        "description": "Respond to a friend request (accept, reject, block).",
        "operationId": "respond_to_friend_request_api_friends_requests__request_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendshipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/friends/": {
      "get": {
        "tags": [
          "Friends"
        ],
        "summary": "Get Friends",
        "description": "Get all friends for the current user.",
        "operationId": "get_friends_api_friends__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "type": "array",
                  "title": "Response Get Friends Api Friends  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/friends/{friend_id}": {
      "delete": {
        "tags": [
          "Friends"
        ],
        "summary": "Remove Friend",
        "description": "Remove a friend.",
        "operationId": "remove_friend_api_friends__friend_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "friend_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Friend Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/calls/": {
      "post": {
        "tags": [
          "Calls"
        ],
        "summary": "Initiate Call",
        "description": "Initiate a call to another user.",
        "operationId": "initiate_call_api_calls__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/calls/{call_id}": {
      "put": {
        "tags": [
          "Calls"
        ],
        "summary": "Update Call Status",
        "description": "Update call status (accept, decline, complete).",
        "operationId": "update_call_status_api_calls__call_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Call Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/calls/history": {
      "get": {
        "tags": [
          "Calls"
        ],
        "summary": "Get Call History",
        "description": "Get call history for the current user.",
        "operationId": "get_call_history_api_calls_history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CallResponse"
                  },
                  "title": "Response Get Call History Api Calls History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CallCreate": {
        "properties": {
          "receiver_id": {
            "type": "integer",
            "title": "Receiver Id"
          }
        },
        "type": "object",
        "required": [
          "receiver_id"
        ],
        "title": "CallCreate"
      },
      "CallResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "caller_id": {
            "type": "integer",
            "title": "Caller Id"
          },
          "receiver_id": {
            "type": "integer",
            "title": "Receiver Id"
          },
          "status": {
            "$ref": "#/components/schemas/CallStatus"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "answered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "caller_id",
          "receiver_id",
          "status",
          "started_at"
        ],
        "title": "CallResponse"
      },
      "CallStatus": {
        "type": "string",
        "enum": [
          "initiated",
          "accepted",
          "declined",
          "missed",
          "completed"
        ],
        "title": "CallStatus"
      },
      "CallUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/CallStatus"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CallUpdate"
      },
      "FriendRequestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "requester_id": {
            "type": "integer",
            "title": "Requester Id"
          },
          "requester_username": {
            "type": "string",
            "title": "Requester Username"
          },
          "requester_avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requester Avatar"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "requester_username",
          "created_at"
        ],
        "title": "FriendRequestResponse"
      },
      "FriendshipCreate": {
        "properties": {
          "addressee_id": {
            "type": "integer",
            "title": "Addressee Id"
          }
        },
        "type": "object",
        "required": [
          "addressee_id"
        ],
        "title": "FriendshipCreate"
      },
      "FriendshipResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "requester_id": {
            "type": "integer",
            "title": "Requester Id"
          },
          "addressee_id": {
            "type": "integer",
            "title": "Addressee Id"
          },
          "status": {
            "$ref": "#/components/schemas/FriendshipStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "requester_id",
          "addressee_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "FriendshipResponse"
      },
      "FriendshipStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "rejected",
          "blocked"
        ],
        "title": "FriendshipStatus"
      },
      "FriendshipUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/FriendshipStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FriendshipUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserResponse": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "status_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Message"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "online_status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Online Status"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "id",
          "timezone",
          "is_active",
          "is_verified",
          "role",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ],
        "title": "UserRole"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}